backend/requirements.txt
fastapi
uvicorn[standard]
openai
SpeechRecognition
python-dotenv
python-multipart
pydantic-settings
langchain-community
aiosqlite
transformers
torch
langchain
langchain-experimental
accelerate
protobuf3
sentencepiece
langchain-openai
langchain_groq